Refactor/99 (#106)
* refactor(dateHandler): formattedDate 추가

* refactor(replyInsertHandler): getReply로 함수명 변경

* refactor: postModal 리펙토링후 에러 제거
