+++ src/components/card/card.jsx
@@ -1,29 +1,43 @@
 import axios from 'axios';
 import { useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import {
+  useDispatch,
+  useSelector,
+} from 'react-redux';
 
 import { postActions } from '../../reducers/post';
-import PostModal from '../postModal/PostModal';
+import PostModal from '../postModal/index';
 import CardItem from './cardItem';
 import './Cards.css';
 
 function Cards() {
   console.log('executes Cards');
   const dispatch = useDispatch();
-  const cardData = useSelector((state) => state.post);
-  const [modalOpacity, setModalOpacity] = useState(0);
+  const cardData = useSelector(
+    (state) => state.post,
+  );
+  const [modalOpacity, setModalOpacity] =
+    useState(0);
   useEffect(() => {
     console.log('rendered Cards');
     axios
-      .get(`${process.env.REACT_APP_HOST}/api/board/getBoard`, {
-        headers: {
-          'X-ACCESS-TOKEN': localStorage.getItem('accessToken'),
+      .get(
+        `${process.env.REACT_APP_HOST}/api/board/getBoard`,
+        {
+          headers: {
+            'X-ACCESS-TOKEN':
+              localStorage.getItem('accessToken'),
+          },
         },
-      })
+      )
       .then((res) => {
-        console.log('postActions.updateIems: Card');
+        console.log(
+          'postActions.updateIems: Card',
+        );
         console.log(res.data);
-        dispatch(postActions.updateItems(res.data));
+        dispatch(
+          postActions.updateItems(res.data),
+        );
       })
       .catch((res) => {
         console.log(res);
