+++ src/components/main/Main.jsx
@@ -1,7 +1,7 @@
 import { useEffect, React } from 'react';
 
 import Cards from '../card/card';
-import Paging from '../pagination/Pagination';
+import Paging from '../Pagination/Pagination';
 
 export default function Main() {
   console.log('executes Main');
